Add ModalHeader
